*delete

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <limits.h>
 
 /**
  * main - prints the first 98 fibonnaci numbers starting with 1 and 0
@@ -8,16 +9,16 @@
  */
 int main(void)
 {
-	double prev_fn = 0;
-	double curr_fn = 1;
-	double fn = 0;
+	long double prev_fn = 0;
+	long double curr_fn = 1;
+	long double fn = 0;
 	int i = 0;
 
 	for (i = 1; i <= 98; ++i)
 	{
 		fn = prev_fn + curr_fn;
 
-		printf("%.0f", fn);
+		printf("%0.0Lf", fn);
 
 		if ((i <= 97))
 			printf(", ");
